Added MIT license
Any chance you'd be willing to add a license declaration to this? (not especially committed to MIT if you'd prefer something else)

I'm hoping to put a fork of this we made (https://github.com/mattsheppard/ftldoc) which turns it into a maven plugin onto maven central, but need to make sure the license stuff is sorted first.